% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_matrix}
\alias{generate_matrix}
\title{Generate a matrix using data and n_out}
\usage{
generate_matrix(x, n_out, remove_first_last = TRUE)
}
\arguments{
\item{x}{Numeric value.
The values that the user wants to aggregate.}

\item{n_out}{Integer.
The number of samples that the user wants to output.}

\item{remove_first_last}{Boolean.
Whether the first and last values are included in obtaining the samples
or raw values are used.}
}
\description{
This is used in the aggregation functions and not be exported.
}
