% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_aggregators}
\alias{list_aggregators}
\title{Show the aggregation functions}
\usage{
list_aggregators()
}
\description{
It displays all the aggregators registered in the package.
No arguments are necessary.
}
\examples{
list_aggregators()
}
