% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTTB.R
\name{LTTB}
\alias{LTTB}
\title{Downsample with the Largest Triangle Three Buckets (LTTB) aggregation method}
\usage{
LTTB(x, y, n_out)
}
\arguments{
\item{x, y}{numeric vectors.}

\item{n_out}{length of the output.
This must be larger than 2 and lesser than the number of the rows of `data`}
}
\value{
named list of `x` and `y`
}
\description{
Downsample with the Largest Triangle Three Buckets (LTTB) aggregation method
}
