% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{mean_aggregator}
\alias{mean_aggregator}
\title{R6 Class for aggregation which returns the mean values}
\format{
An \code{R6::R6Class} object
}
\description{
The mean value for each interval is calculated.
}
\section{Super classes}{
\code{\link[shinyHugePlot:abstract_aggregator]{shinyHugePlot::abstract_aggregator}} -> \code{\link[shinyHugePlot:custom_stat_aggregator]{shinyHugePlot::custom_stat_aggregator}} -> \code{mean_aggregator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mean_aggregator-new}{\code{mean_aggregator$new()}}
\item \href{#method-mean_aggregator-clone}{\code{mean_aggregator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shinyHugePlot" data-topic="abstract_aggregator" data-id="aggregate"><a href='../../shinyHugePlot/html/abstract_aggregator.html#method-abstract_aggregator-aggregate'><code>shinyHugePlot::abstract_aggregator$aggregate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean_aggregator-new"></a>}}
\if{latex}{\out{\hypertarget{method-mean_aggregator-new}{}}}
\subsection{Method \code{new()}}{
Constructor of the Aggregator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mean_aggregator$new(interleave_gaps = FALSE, nan_position = "end")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interleave_gaps, nan_position}}{Arguments pass to the constructor of
the \code{abstract_aggregator} class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean_aggregator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mean_aggregator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mean_aggregator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
