% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nanosecond_to_label}
\alias{nanosecond_to_label}
\title{Generate an appropriate time-based label using nanotime value}
\usage{
nanosecond_to_label(ns)
}
\arguments{
\item{ns}{Numeric value(s) representing nano second.}
}
\value{
Character.
}
\description{
Nano-second time is converted to an appropriate label.
e.g., 1e9 nano seconds will be converted to \code{"1.0s"}
}
