% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nanotime_to_plotlytime}
\alias{nanotime_to_plotlytime}
\title{Conversion of nanotime to plotly-style time}
\usage{
nanotime_to_plotlytime(time, tz)
}
\arguments{
\item{time}{nanotime class time.}

\item{tz}{time zone (e.g. "Asia/Tokyo")}
}
\value{
local time (strings) with the format of %Y-%m-%d %H:%M:%E9S
}
\description{
Plotly does not accept 64-bit nanotime format,
so the conversion is necessary to import the time to plotly.
}
