% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_fluct.R
\docType{data}
\name{noise_fluct}
\alias{noise_fluct}
\title{Time-series fluctuations in sound level}
\format{
A data frame with 32,001 rows and 4 columns:
\describe{
  \item{time}{time}
  \item{f500, f1000, f2000}{Octave-band sound levels whose center frequencies are 500, 1000 and 2000 Hz.}
}
}
\usage{
noise_fluct
}
\description{
Results of the measurement of the sound level,
where peaks due to road traffic are observed.
}
\author{
Junta Tagusari \email{j.tagusari@eng.hokudai.ac.jp}
}
\keyword{noise}
\keyword{sound}
\keyword{time-series}
\keyword{traffic}
