% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotlytime_to_nanotime}
\alias{plotlytime_to_nanotime}
\title{Conversion of plotly-style time to nanotime}
\usage{
plotlytime_to_nanotime(time, tz)
}
\arguments{
\item{time}{local time (strings) with the format of
`\%Y-\%m-\%d \%H:\%M:\%S.s`}

\item{tz}{time zone (e.g. "Asia/Tokyo")}
}
\value{
nanotime
}
\description{
Datetime obtained from plotly are converted to 64-bit nanotime format.
}
