% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyHugePlot.R
\docType{package}
\name{shinyHugePlot}
\alias{shinyHugePlot}
\title{shinyHugePlot}
\description{
An interactive plot for data with a large sample size using \code{shiny}
and\code{plotly} can be obtained.
For an easy application, see \code{shiny_hugeplot} function.
For a manual application, see \code{downsampler} class.
}
