% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nanosecond_to_label}
\alias{nanosecond_to_label}
\title{Generate an time-based aggregation label using nanotime interval}
\usage{
nanosecond_to_label(ns)
}
\arguments{
\item{ns}{Numeric value(s) representing nano second.}
}
\value{
Character.
}
\description{
Nano-second time is converted to an appropriate aggregation label,
which is used for the legend of the aggregated series.
e.g., 1e9 nano seconds will be converted to \code{"1.0s"}
}
