% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_fluct.R
\docType{data}
\name{noise_fluct}
\alias{noise_fluct}
\title{Time-series fluctuations in sound level}
\description{
Results of the measurement of the sound level,
where peaks due to road traffic are observed.
}
\author{
Junta Tagusari \email{j.tagusari@eng.hokudai.ac.jp}
}
\keyword{noise}
\keyword{sound}
\keyword{time-series}
\keyword{traffic}
